Implement basic \ref dynamic completion
Use \label to populate \ref completion.
This is a proof-of-concept using the parser from document_structure.

A bit of refactoring.

It's alive guys!

	Completion for the \ref command is now available for the opened documents.
	Improvements to be added:
	 - handling the different projects, to provide relevant completion choices.
	 - maybe use the opening of a file to call the parser, so the completion is immediatly available. This is the problem for using the parser of document-structure, which is called when creating the tree view.

Differentiating the projects.

Now, the completion choices are related only to the parsed files of the same parent directory than the one of the current document.
Still, the parsing needs to be called for all the tex files of the directory, when opening a document. Thus, the code can be left untouched, because saving an opened document will update the small choices that need to be changed.

Efficiency improvements.

The completion choices for the \ref command are now updated only if neccessary (there is still a case where it could be updated later, but we'll stay with this version for now).
Thus there is a big gain of efficiency.

Plus, the documents' document_structure is now parsed when opening the document, thus providing the completion when loading a file.

Still, we need to parse all the documents in the same project than the one we open.

Also, it would be nice to parse the document when switching from one document_tab to another.

Switch completion context when switching documents

The completion provider is now notified when the user switches tab, and changes the completion context accordingly
It provides \ref completion only for \labels sourced by documents opened AND in the same folder as the currently focused document.