Differentiating the projects.
Now, the completion choices are related only to the parsed files of the same parent directory than the one of the current document.
Still, the parsing needs to be called for all the tex files of the directory, when opening a document. Thus, the code can be left untouched, because saving an opened document will update the small choices that need to be changed.