Efficiency improvements.
The completion choices for the \ref command are now updated only if neccessary (there is still a case where it could be updated later, but we'll stay with this version for now).
Thus there is a big gain of efficiency.

Plus, the documents' document_structure is now parsed when opening the document, thus providing the completion when loading a file.

Still, we need to parse all the documents in the same project than the one we open.

Also, it would be nice to parse the document when switching from one document_tab to another.