Add support for parsing labels in non opened docs.
The labels are now parsed on the loading of a document. In this case, the parsing is also made for all the tex files in the same directory as the opened document.

The labels parsing is now detached from document_structure, thus all the previous changes in document_structure should now be reverted.

The label parser is now in document.vala, and is an extract of the existing parser in document_structure.

One new error though: upon closing a document in the ui, a gtk_is_visible error is raised.